requirements.txt
Django==4.0.1
djangorestframework==3.12.4
psycopg2==2.9.2
pytz==2021.3
requests==2.27.1
